Montrer tout le volume sur la carte d'une passe
La part de clusters par bloc était arrondie à l'entier inférieur, et le
reste de la division s'entassait dans la dernière case : en plein écran,
le dernier cinquième de « Secrétariat, 1996 » tenait dans un seul bloc.
Chaque bloc reçoit désormais sa part exacte, à un cluster près.

+++ Sources/Model/DiskLibraryModel.swift
@@ -171,9 +171,10 @@ final class DiskLibraryModel: ObservableObject {
 
     // MARK: - Lecture pour l'affichage
 
-    var clustersPerCell: Int {
+    /// Le découpage de `GeneratedDisk.cells`, fraction comprise.
+    var clustersPerCell: Double {
         guard let disk = state.disk else { return 1 }
-        return max(Int(disk.clusterCount) / grid.cellCount, 1)
+        return max(Double(disk.clusterCount) / Double(grid.cellCount), 1)
     }
 
     /// Catégories réellement présentes, pour ne légender que ce qu'on voit.
